requirements.txt
pandas 
scikit-learn
python-dotenv
yara-python
PyGithub
numpy
os
subprocess
shutil
stat
pickle
sys
